+++ client/src/thin_client.rs
@@ -8,6 +8,7 @@ use bincode::{serialize_into, serialized_size};
 use log::*;
 use solana_sdk::account::Account;
 use solana_sdk::client::{AsyncClient, Client, SyncClient};
+use solana_sdk::clock::MAX_PROCESSING_AGE;
 use solana_sdk::fee_calculator::FeeCalculator;
 use solana_sdk::hash::Hash;
 use solana_sdk::instruction::Instruction;
@@ -16,7 +17,7 @@ use solana_sdk::packet::PACKET_DATA_SIZE;
 use solana_sdk::pubkey::Pubkey;
 use solana_sdk::signature::{Keypair, KeypairUtil, Signature};
 use solana_sdk::system_instruction;
-use solana_sdk::timing::{duration_as_ms, MAX_PROCESSING_AGE};
+use solana_sdk::timing::duration_as_ms;
 use solana_sdk::transaction::{self, Transaction};
 use solana_sdk::transport::Result as TransportResult;
 use std::io;
