+++ core/src/blocktree_processor.rs
@@ -7,9 +7,10 @@ use rayon::ThreadPool;
 use solana_metrics::{datapoint, datapoint_error, inc_new_counter_debug};
 use solana_runtime::bank::Bank;
 use solana_runtime::locked_accounts_results::LockedAccountsResults;
+use solana_sdk::clock::{Slot, MAX_RECENT_BLOCKHASHES};
 use solana_sdk::genesis_block::GenesisBlock;
 use solana_sdk::hash::Hash;
-use solana_sdk::timing::{duration_as_ms, Slot, MAX_RECENT_BLOCKHASHES};
+use solana_sdk::timing::duration_as_ms;
 use solana_sdk::transaction::Result;
 use std::result;
 use std::sync::Arc;
