+++ core/src/chacha_cuda.rs
@@ -120,9 +120,9 @@ mod tests {
     use crate::chacha_cuda::chacha_cbc_encrypt_file_many_keys;
     use crate::entry::make_tiny_test_entries;
     use crate::replicator::sample_file;
+    use solana_sdk::clock::DEFAULT_SLOTS_PER_SEGMENT;
     use solana_sdk::hash::Hash;
     use solana_sdk::signature::{Keypair, KeypairUtil};
-    use solana_sdk::timing::DEFAULT_SLOTS_PER_SEGMENT;
     use std::fs::{remove_dir_all, remove_file};
     use std::path::Path;
     use std::sync::Arc;
