+++ core/src/replicator.rs
@@ -27,10 +27,11 @@ use solana_ed25519_dalek as ed25519_dalek;
 use solana_netutil::bind_in_range;
 use solana_sdk::account_utils::State;
 use solana_sdk::client::{AsyncClient, SyncClient};
+use solana_sdk::clock::{get_complete_segment_from_slot, get_segment_from_slot};
 use solana_sdk::hash::{Hash, Hasher};
 use solana_sdk::message::Message;
 use solana_sdk::signature::{Keypair, KeypairUtil, Signature};
-use solana_sdk::timing::{get_complete_segment_from_slot, get_segment_from_slot, timestamp};
+use solana_sdk::timing::timestamp;
 use solana_sdk::transaction::Transaction;
 use solana_sdk::transport::TransportError;
 use solana_storage_api::storage_contract::StorageContract;
