+++ core/src/storage_stage.rs
@@ -15,12 +15,12 @@ use solana_runtime::bank::Bank;
 use solana_runtime::storage_utils::replicator_accounts;
 use solana_sdk::account::Account;
 use solana_sdk::account_utils::State;
+use solana_sdk::clock::get_segment_from_slot;
 use solana_sdk::hash::Hash;
 use solana_sdk::instruction::Instruction;
 use solana_sdk::message::Message;
 use solana_sdk::pubkey::Pubkey;
 use solana_sdk::signature::{Keypair, KeypairUtil, Signature};
-use solana_sdk::timing::get_segment_from_slot;
 use solana_sdk::transaction::Transaction;
 use solana_storage_api::storage_contract::{Proof, ProofStatus, StorageContract};
 use solana_storage_api::storage_instruction;
@@ -638,10 +638,10 @@ mod tests {
     use crate::{blocktree_processor, entry};
     use rayon::prelude::*;
     use solana_runtime::bank::Bank;
+    use solana_sdk::clock::DEFAULT_TICKS_PER_SLOT;
     use solana_sdk::hash::{Hash, Hasher};
     use solana_sdk::pubkey::Pubkey;
     use solana_sdk::signature::{Keypair, KeypairUtil};
-    use solana_sdk::timing::DEFAULT_TICKS_PER_SLOT;
     use std::cmp::{max, min};
     use std::fs::remove_dir_all;
     use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
