+++ runtime/src/stakes.rs
@@ -1,9 +1,9 @@
 //! Stakes serve as a cache of stake and vote accounts to derive
 //! node stakes
 use solana_sdk::account::Account;
+use solana_sdk::clock::Epoch;
 use solana_sdk::pubkey::Pubkey;
 use solana_sdk::sysvar::stake_history::StakeHistory;
-use solana_sdk::timing::Epoch;
 use solana_stake_api::stake_state::{new_stake_history_entry, StakeState};
 use solana_vote_api::vote_state::VoteState;
 use std::collections::HashMap;
