+++ runtime/src/status_cache.rs
@@ -1,13 +1,13 @@
 use log::*;
 use rand::{thread_rng, Rng};
 use serde::Serialize;
+use solana_sdk::clock::{Slot, MAX_HASH_AGE_IN_SECONDS};
 use solana_sdk::hash::Hash;
 use solana_sdk::signature::Signature;
-use solana_sdk::timing::Slot;
 use std::collections::{HashMap, HashSet};
 use std::sync::{Arc, Mutex};
 
-pub const MAX_CACHE_ENTRIES: usize = solana_sdk::timing::MAX_HASH_AGE_IN_SECONDS;
+pub const MAX_CACHE_ENTRIES: usize = MAX_HASH_AGE_IN_SECONDS;
 const CACHED_SIGNATURE_SIZE: usize = 20;
 
 // Store forks in a single chunk of memory to avoid another lookup.
