+++ validator/src/main.rs
@@ -12,9 +12,9 @@ use solana_core::ledger_cleanup_service::DEFAULT_MAX_LEDGER_SLOTS;
 use solana_core::service::Service;
 use solana_core::socketaddr;
 use solana_core::validator::{Validator, ValidatorConfig};
+use solana_sdk::clock::Slot;
 use solana_sdk::hash::Hash;
 use solana_sdk::signature::{read_keypair, Keypair, KeypairUtil};
-use solana_sdk::timing::Slot;
 use std::fs::{self, File};
 use std::io::{self, Read};
 use std::net::SocketAddr;
